+++ Management/Models/Annotations/ReportServer.cs
@@ -1,0 +1,90 @@
+﻿using DisplayMonkey.Language;
+using System;
+using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
+using System.Linq;
+using System.Web;
+
+namespace DisplayMonkey.Models
+{
+    [
+        MetadataType(typeof(ReportServer.Annotations))
+    ]
+    public partial class ReportServer
+    {
+        internal class Annotations
+        {
+            [
+                Display(ResourceType = typeof(Resources), Name = "ID"),
+            ]
+            public int ServerId { get; set; }
+
+            [
+                Display(ResourceType = typeof(Resources), Name = "Name"),
+                Required(ErrorMessageResourceType = typeof(Resources), ErrorMessageResourceName = "NameRequired"),
+                StringLength(100, ErrorMessageResourceType = typeof(Resources), ErrorMessageResourceName = "MaxLengthExceeded"),
+            ]
+            public string Name { get; set; }
+
+            [
+                Display(ResourceType = typeof(Resources), Name = "BaseURL"),
+                Required(ErrorMessageResourceType = typeof(Resources), ErrorMessageResourceName = "BaseURLRequired"),
+            ]
+            public string BaseUrl { get; set; }
+
+            [
+                Display(ResourceType = typeof(Resources), Name = "User"),
+            ]
+            public string User { get; set; }
+
+            [
+                Display(ResourceType = typeof(Resources), Name = "Password"),
+            ]
+            public byte[] Password { get; set; }
+
+            [
+                Display(ResourceType = typeof(Resources), Name = "Domain"),
+            ]
+            public string Domain { get; set; }
+
+            [
+                Display(ResourceType = typeof(Resources), Name = "Reports"),
+            ]
+            public virtual ICollection<Report> Reports { get; set; }
+        }
+
+        [
+            Display(ResourceType = typeof(Resources), Name = "Password"),
+            DataType(DataType.Password),
+            NotMapped,
+        ]
+        public string PasswordUnmasked
+        {
+            get
+            {
+                //if (this.Password == null)
+                //    return null;
+                //else
+                //    return RsaUtil.Decrypt(this.Password);
+                return Constants.PasswordMask;
+            }
+
+            set
+            {
+                PasswordSet = true;
+                if (value == null)
+                    this.Password = null;
+                else if (value != Constants.PasswordMask)
+                    this.Password = RsaUtil.Encrypt(value);
+                else
+                    PasswordSet = false;
+            }
+        }
+
+        [
+            NotMapped,
+        ]
+        public bool PasswordSet { get; private set; }
+    }
+}
