+++ Management/Views/Panel/Index.cshtml
@@ -1,4 +1,5 @@
-﻿@using DisplayMonkey.Language
+﻿@using DisplayMonkey.Models
+@using DisplayMonkey.Language
 
 @model IEnumerable<DisplayMonkey.Models.Panel>
 
@@ -25,10 +26,10 @@
         <th>
         </th>
         <th class="nw">
-            @Html.DisplayNameFor(model => model.Canvas)
+            @Resources.Canvas
         </th>
         <th>
-            @Html.DisplayNameFor(model => model.Canvas.FullScreen.Panel)
+            @Resources.Panel
         </th>
         <th class="nw">
             @Resources.Options
@@ -38,7 +39,7 @@
 
 @foreach (var item in Model)
 {
-    DisplayMonkey.Models.FullScreen fs = item.Canvas.FullScreen;
+    FullScreen fs = item.Canvas.FullScreen;
     
     <tr>
         <td>
