+++ Management/Views/Panel/_Details.cshtml
@@ -1,9 +1,10 @@
-﻿@using DisplayMonkey.Language
+﻿@using DisplayMonkey.Models
+@using DisplayMonkey.Language
 
 @model DisplayMonkey.Models.Panel
 
 @{
-    DisplayMonkey.Models.FullScreen fs = Model.Canvas.FullScreen;
+    FullScreen fs = Model.Canvas.FullScreen;
 }
 
 <fieldset>
